requirements-dev.txt
asgiref==3.2.7
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
Pillow==7.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
